pyproject.toml
[project]
name = "portable-llm"
version = "0.1.0"
dependencies = [
    "llama-cpp-python>=0.2.26",
    "pyyaml>=6.0",
    "requests>=2.0.0",
    "huggingface-hub>=0.20.3",
]
requires-python = ">=3.8"
